backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
werkzeug>=3.0.0
requests>=2.28.0
python-dateutil>=2.8.0
pytest>=7.4.0
gunicorn>=21.0.0
pypdf>=5.0.0
python-docx>=1.1.0
